+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ## [Unreleased]
 ### Added
+- Add `Identification.forReviewAt`
 - Add `EnvelopeRecipient.channelForNotifications`
 - Add `EnvelopeTemplateRecipient.channelForNotifications`
 - Add `DeactivateAccount` endpoint