+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 ## [Unreleased]
 ### Added
 - Add multiple Identifications endpoints
+- Add `IdentificationInspection.claims`
 - Add `Identificaiton.cancelledAt`
 - Add `Identificaiton.discardedAt`
 - Add `Identification.result` and its sub-entities