+++ CHANGELOG.md
@@ -7,7 +7,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 ## [Unreleased]
 ### Added
 - Add `SentEnvelopeReport.bankIdSignCount`
-- Add `SentEnvelopeReport.identifyAmlCount`
+- Add `SentEnvelopeReport.bankIdIdentifyAmlCount`
 - Add `IdentificationEndpoint.delete`
 - Add `IdentifyScenarioVersion.expireAfterHours`
 - Add `IdentifyScenarioVersion.discardCompletedAfterDays`