+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
-## [Unreleased]
+## [2.3.0] - 2024-03-28
 ### Added
 - Add `SentEnvelopeReport.bankIdSignCount`
 - Add `SentEnvelopeReport.bankIdIdentifyAmlCount`